Refactor unified turn loop into submodules
