+++ src/agent/runloop/unified/turn/mod.rs
@@ -1,0 +1,9 @@
+mod config_modal;
+mod harmony;
+mod session;
+mod workspace;
+
+pub(crate) use session::run_single_agent_loop_unified;
+
+#[cfg(test)]
+pub(crate) use harmony::strip_harmony_syntax;
